added calendar from JTAppleCalendar

+++ travel-weather/Data Models/Day.swift
@@ -21,7 +21,7 @@ class Day {
     var date: Date
     var highTemp: Int?
     var lowTemp: Int?
-    var latLong: (Float, Float)?
+    var latLong: (Double, Double)?
     var weatherSummary: WeatherSummary?
     
     //MARK:- Computed Properties
